% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_package}
\alias{check_package}
\title{Help-Function}
\usage{
check_package(packages)
}
\arguments{
\item{packages}{A vector of packages}
}
\description{
This function checks if a a vector of packages named each as a string are installed. Every package that is installed will be loaded and all not yet
installed packages will be installed and loaded.
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}
