% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{data_cut}
\alias{data_cut}
\title{Help-Function}
\usage{
data_cut(data, var, value, prob = 0.15)
}
\arguments{
\item{data}{The data to be analysed}

\item{var}{The variable upon the subpopulation should depend on}

\item{value}{The value of the given variable}

\item{prob}{What percentage of the data over and under the given limit should be returned (default is 0.15)}
}
\description{
This function returns a subpopulation of the given data that lies around a given value of a given variable
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

