% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{factor_check}
\alias{factor_check}
\title{Help-Function}
\usage{
factor_check(data)
}
\arguments{
\item{data}{The data to be analysed}
}
\description{
This function tells you what variables of a given dataset are factorial variables
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}
