% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{factor_melt}
\alias{factor_melt}
\title{Help-Function}
\usage{
factor_melt(factor_leveln1, factor_leveln2, factor1)
}
\arguments{
\item{factor_leveln1}{The number of the first factor level}

\item{factor_leveln2}{The number of the second factor level}

\item{factor1}{The factor variable itself}
}
\description{
This function melts two factor levels into one factor level.
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}
