% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{faktor}
\alias{faktor}
\title{Help-Function}
\usage{
faktor(data)
}
\arguments{
\item{data}{The data to be analysed}
}
\description{
This function turns all factorial variable in a dataset into metric variables and returns a logical vector what variable were factor-variables.
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

