% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{faktor2}
\alias{faktor2}
\title{Help-Function}
\usage{
faktor2(data, n)
}
\arguments{
\item{data}{A data.frame object}

\item{n}{A numeric value indicating from what number of different values a variable is seen as categorical variables, all variables that have more different values than n are being treated as metric values}
}
\description{
This function factorizes a variables with more than n different values.
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

