% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{fraction_variable}
\alias{fraction_variable}
\title{Help-Function}
\usage{
fraction_variable(variable, a = 100)
}
\arguments{
\item{variable}{A numeric vector}

\item{a}{How much more freedom do you need with the plot-limits? The lower the more freedomn}
}
\description{
This function is used in the Metric_Single_Variable_Analysis-function to make sure
that you always have a bit more freedom with decision what data to plot that what data to actually use.
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

