% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{polynomfit}
\alias{polynomfit}
\title{Help-Function}
\usage{
polynomfit(d, x, y)
}
\arguments{
\item{d}{The degree of the fit}

\item{x}{The independent metric variable}

\item{y}{The dependent metric variable}
}
\description{
This function calculates a polynomial-fit
}
\note{
The output also features a prediction of the dependent variable in the interval from min(x) to max(x)
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

