% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{position}
\alias{position}
\title{Help-Function}
\usage{
position(x, y, data)
}
\arguments{
\item{x}{the x-value of the click}

\item{y}{the y-value of the click}

\item{data}{The data to be analysed}
}
\description{
This function calculates onto what scatterplot in the scatterplot-matrix the user clicked. (Function used in the function Scatterplot_Matrix)
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

