% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{position2}
\alias{position2}
\title{Help-Function}
\usage{
position2(x, n)
}
\arguments{
\item{x}{A numeric value (the x value of the point where the user clicked)}

\item{n}{A numeric value (how many boxplots are there)}
}
\description{
This function is used in the boxplot application to find out onto what boxplot a used clicked.
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

