% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{position3}
\alias{position3}
\title{Help-Function}
\usage{
position3(x)
}
\arguments{
\item{x}{A numeric value (the x value of the point where the user clicked)}
}
\description{
This function is used in the Stacked_Barplot application to find out onto what barplot a used clicked.
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

