% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{position4}
\alias{position4}
\title{Help-Function}
\usage{
position4(y)
}
\arguments{
\item{y}{The y-value of a click}
}
\description{
This function is used on the Timeseries-functions and recognizes from the observed y variable
of a click in what the used is clicking.
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

