% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{position5}
\alias{position5}
\title{Help-Function}
\usage{
position5(x, ymin, ymax, ts_w)
}
\arguments{
\item{x}{A numeric value ether 1,2,3 or 4. There are four plots shown underneath each other, what plot should be zoomed in?}

\item{ymin}{The ymin value of the 'brush'-interaction}

\item{ymax}{The ymax value of the 'brush'-interaction}

\item{ts_w}{The analysed time series}
}
\description{
This function is used on the Timeseries-functions and recognizes the y-limits of the 'brush'-interaction in the applications.
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

