% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{position6}
\alias{position6}
\title{Help-Function}
\usage{
position6(xmin, xmax, beginning, ending)
}
\arguments{
\item{xmin}{The ymin value of the 'brush'-interaction}

\item{xmax}{The ymax value of the 'brush'-interaction}

\item{beginning}{The beginning of the analysed time series}

\item{ending}{The ending of the analysed time series}
}
\description{
This function is used on the Timeseries-functions and recognizes the x-limits of the 'brush'-interaction in the applications.
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

