% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{range01}
\alias{range01}
\title{Help-Function}
\usage{
range01(x)
}
\arguments{
\item{x}{A ordered numeric vector}
}
\description{
This function scales the values of a ordered vector from 0 to 1
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

