% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{round_check}
\alias{round_check}
\title{Help-Function}
\usage{
round_check(x)
}
\arguments{
\item{x}{The vector to be checked}
}
\description{
This function checks if a numeric vector only includes whole numbers
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

