% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatterplotly_Matrix.R
\name{Scatterplot_Matrix}
\alias{Scatterplot_Matrix}
\title{Interactive Scatterplot-Matrix}
\usage{
Scatterplot_Matrix(data, metr_data = F, width = c(400, 700, 400),
  height = c(500, 700, 500))
}
\arguments{
\item{data}{A data.frame object that is to be analyzed (all categorical variables with be transformed to metrical variables)}

\item{metr_data}{A logical value, indicating whether a table showing the metric and categorical analogies of levels should be shown}

\item{width}{A three-dimensional numeric value indicating the width of the three shown plots}

\item{height}{A three-dimensional numeric value indicating the height of the three shown plots}
}
\description{
Takes any data and creates an interactive Scatterplot-Matrix with multiple features.
}
\details{
At first one has to choose the variables to be plotted in the scatterplot matrix. Once the user has choosen at least two variables (the minimum of a scatterplot),
those two plots are also shown on each side of the Scatterplot-Matrix.
The user though can decide what scatterplots out of the Matrix should be zoom in on each side. A conventional click on one scatterplot from the Scatterplot-Matrix triggers a bigger scatter plot on the left-hand side,
while you can control the scatterplot on the right-hand side with a double click on the desired scatterplot in the Scatterplot-Matrix.
On each of the small scatterplots one can select certain points with the brush option, implemented in the R-packge Shiny. Once you have chosen a cloud of data points
and performed a normal click, the chosen points with be colored red in all available scatterplots. If you want to see the chosen data you have to
press the "show data"-Button. A linear regression line and a loess curve can also be plotted by pressing the fitting button. In the case of a pressed "Regression"- or "Smooth"-Button.
If groups have been definied by brushing actions, you can also plot the linear and loess regression by group once you pressed the Button "By Group".
}
\examples{
data=iris
if (interactive()) {
  Scatterplot_Matrix(data,metr_data = T)
  }
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

