% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing_Analysis.R
\name{Smoothing_Analysis}
\alias{Smoothing_Analysis}
\title{Interactive smoothing with two metric variables}
\usage{
Smoothing_Analysis(data, n = 10, height = 500)
}
\arguments{
\item{data}{A data.frame object that is to be analyzed (only metric variables are shown)}

\item{n}{A numeric value indicating from what number of different values a variable is seen as categorical, all variables that have more different values than n are being treated as metric values}

\item{height}{A numeric value indicating the height of the shown scatterplot}
}
\description{
Takes any metric data and plots a two-dimensional scatterplot, you can fit a polynomial or smoothing spline with a confidence band.
}
\details{
On the sidebar you can chose two metric variables from the data set to be plotted in a scatterplot. If you wish you can also transform any of the used variables (options are: identity, log, exp, sin,cos).
On the right side you see the R^2 values of the current model. To fit a polynomial model chose the 'Polynomial'-option in the type of regression.With a numeric input you can change the degree of the fit.
Another option is to fit b-Splines, you can chose them as the type of regression. And with two numerical inputs you can set the degree and knots of the fit. The knots
arguments setts the amount of splits in the data and the degree argument the degree of each of these splits. The last optional type of regression is smoothing.
At first you can set the amount of knots and decide weather you want to penalize the fit. You can also use different kind of smoothing basis you want to use. There are four different possible smoothing bases:
thing plate regression splines, cubic regression splines, p-splines and duchon splines.
If you want to penalize the fit check the box 'Penalize', there are now two more variables
that you can change. You can change the order of penalty and a value called lambda telling the model how important the penalty term should be (the higher the more important). Y
You can also change the family and link-Function of the model. For further reading on these options see the documentation in the stats-Package (\code{\link{family}} and \code{\link{make.link}}).
}
\examples{
data=iris
if (interactive()) {
  Smoothing_Analysis(data,n=10,height = 500)
  }
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

