% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{table_transform}
\alias{table_transform}
\title{Help-Function}
\usage{
table_transform(table, click)
}
\arguments{
\item{table}{Table to be transformed}

\item{click}{The number of the level the used clicked on in the application}
}
\description{
This function is used in the Stacked_Barplot application and transforms a given table so that one can specify the colour in the barplot.
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

