% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timeseries.R
\name{Timeseries}
\alias{Timeseries}
\title{Interactive Seasonal Decomposition of Time Series by Loess}
\usage{
Timeseries(data, height = c(800, 400), n = 10)
}
\arguments{
\item{data}{A data.frame object that is to be analyzed (all categorical variables with be transformed to metrical variables)}

\item{height}{A two-dimensional numeric value indicating the height of the decomposition and the heigth of the zoom-graph}

\item{n}{A numeric value indicating from what number of different values a variable is seen as categorical variables, all variables that have more different values than n are being treated as metric values}
}
\description{
Takes any metric data with time series and plots a seasional decomposition by Loess.
}
\details{
At first you can chose what metric variable should be composed into a seasonal, trend and remainder part. Then you can say from what year to what year the data was observed
and the frequency of observations per year. For monthly observed data this parameter should be 12. You can also fit a robust fitting of the loess procedure by pressing the button
called 'Robust fitting'. The default for the window of the seasonal loess window is periodic, meaning as long as one period. This means that the window is exactly as broad as one period
number of the frequency in observations. If you brush over any part of the decomposition you see a zoom plot in the sidebar panel of the shaded area.
}
\examples{

date1=as.Date(paste0(2000,"-1-1"))
date2=as.Date(paste0(2010,"-1-1"))
if (interactive()) {
  data=get.hist.quote(instrument= "BMW.DE",start = date1, end = date2 ,
                      quote=c("Open", "High", "Low", "Close",
                              "AdjClose","Volume"), provider = "yahoo",
                      compression = "m",retclass="zoo",quiet = T)
  data=as.data.frame(data)

  Timeseries(data=data)
}
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

