% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_Series_Stat.R
\name{Timeseries_Stat}
\alias{Timeseries_Stat}
\title{Interactive Seasonal Decomposition of Time Series by Loess example application (Type 2)}
\usage{
Timeseries_Stat(height = 150, width = 800)
}
\arguments{
\item{height}{The height of all plots}

\item{width}{The width of all plots}
}
\description{
For this example a working internet connection is needed.
The user can chose out of a range of stocks and indicators from  on yahoo finance what data from what time (in years) should be downloaded as a time series.
Then a Seasonal Decomposition of Time Series by Loess is performed on the given time series.
}
\details{
If you brush over any part of the decomposition you see a zoom plot in the sidebar panel of the shaded area.
}
\examples{
height = c(400,800)
if (interactive()) {
  Timeseries_Stat(height =height )
}
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}
