% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Timeseries_Stat.R
\name{Timeseries_Stat1}
\alias{Timeseries_Stat1}
\title{Interactive Seasonal Decomposition of Time Series by Loess example application (Type 1)}
\usage{
Timeseries_Stat1(height = c(800, 400))
}
\arguments{
\item{height}{A two-dimensional numeric value indicating the height of the decomposition and the heigth of the zoom-graph}
}
\description{
For this example a working internet connection is needed.
The user can type the name of any stock on yahoo finance to be downloaded as time series and then plot a Seasonal Decomposition of Time Series by Loess
}
\details{
If you brush over any part of the decomposition you see a zoom plot in the sidebar panel of the shaded area.
}
\examples{
height = c(400,800)
if (interactive()) {
  Timeseries_Stat1(height =height )
  Timeseries_Stat2(height =height)
}
}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}

