% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatterplot_3d.R
\name{Scatterplot_3d}
\alias{Scatterplot_3d}
\title{Interactive three-dimensional scatterlpot}
\usage{
Scatterplot_3d(data, n = 10, height = c(1500, 500))
}
\arguments{
\item{data}{A data.frame object that is to be analyzed (only metric variables are shown)}

\item{n}{A numeric value indicating the limit from what number of different values a variable is seen as categorical variable, all variables that have more than n different values are being treated as metric values}

\item{height}{A two-dimensional numeric vector indicating the height of the shown three-dimensional and two-dimensional scatterplots}
}
\description{
Takes any metric data and plots a three-dimensional scatterplot and a two-dimensional scatterplot.
}
\details{
This function produces an application that lets you plot a interactive three-dimensional scatterplot, you can decide what metric variables should be plotted. Next to the
three-dimensional scatterplot there is a two-dimensional scatterplot of the first two dimensions of the three-dimensional scatterplot (1.Variable and 2. Variable). You can change the plotted angle
interactively by clicking and brushing over the shown three-dimensional scatterplot with the cursor. You can press the little 'Play'-Button and it will automatically rotate the plot once.
There is also a feature added in this application, which lets you
condition both plots on a fourth variable. To enter this mode press the button called 'Condition plot only values conditioned on the value of a fourth variable. You can now set a value for the fourth variable with a slider.
Only 30% of the data around that value will now be plotted.
}
\examples{
if (interactive()) {
  Scatterplot_3d(mtcars)
}

}
\author{
Cornelius Fritz <cornelius.fritz@campus.lmu.de>
}
